HW2/requirements.txt
# requirements.txt
# PyTorch CUDA 12.8 wheels
--extra-index-url https://download.pytorch.org/whl/cu128

absl-py==2.3.1
accelerate==0.34.2
af3==3.0.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astropy==6.1.7
astropy-iers-data==0.2025.10.20.0.39.8
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
audiobox-aesthetics==0.0.4
audiocraft==1.4.0a2
audiomentations==0.37.0
audioread==3.0.1
av==11.0.0
babel==2.17.0
beartype==0.22.2
beautifulsoup4==4.14.2
bert-score==0.3.13
bitsandbytes==0.48.1
bleach==6.2.0
blis==0.7.11
braceexpand==0.1.7
catalogue==2.0.10
certifi==2025.10.5
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.1
colorlog==6.10.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.2
datasets==2.16.1
debugpy==1.8.17
decorator==5.2.1
decord==0.6.0
deepspeed==0.15.4
defusedxml==0.7.1
demucs==4.0.0
diffq==0.2.4
diffusers==0.30.3
dill==0.3.7
distlib==0.4.0
distro==1.9.0
docopt==0.6.2
docstring-parser==0.17.0
dora-search==0.1.12
easydict==1.13
einops==0.6.1
einops-exts==0.0.4
encodec==0.1.1
et-xmlfile==2.0.0
evaluate==0.4.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.119.0
fastjsonschema==2.21.2
ffmpeg==1.4
ffmpy==0.6.3
filelock==3.19.1
fire==0.7.1
flashy==0.0.2
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.12
gitpython==3.1.45
gradio==3.36.1
gradio-client==0.2.9
h11==0.16.0
h5py==3.15.1
hf-transfer==0.1.9
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
identify==2.6.15
idna==3.11
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.3.0
iopath==0.1.10
ipykernel==7.0.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.9
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kaldiio==2.18.1
kiwisolver==1.4.9
laion-clap==1.1.7
lameenc==1.8.1
langcodes==3.5.0
language-data==1.3.0
lark==1.3.0
latex2mathml==3.78.1
lazy-loader==0.3
levenshtein==0.27.1
linkify-it-py==2.0.3
loguru==0.7.3
lpmc==0.0.0
madmom==0.16.1
marisa-trie==1.3.1
markdown-it-py==2.0.3
markdown2==2.5.4
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mido==1.3.3
mistune==3.1.4
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.15
muq==0.1.0
murmurhash==1.0.13
music2latent==0.1.6
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
nltk==3.8.1
nnaudio==0.3.3
nodeenv==1.9.1
notebook==7.4.7
notebook-shim==0.2.4
num2words==0.5.14
numba==0.58.1
numpy==1.26.4
numpy-minmax==0.3.1
numpy-rms==0.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==1.8.0
openai-whisper==20250625
opencv-python-headless==4.8.0.76
openpyxl==3.1.2
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.5
peft==0.13.0
pesq==0.0.4
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
portalocker==3.2.0
pre-commit==4.3.0
preshed==3.0.10
pretty-midi==0.2.11
primepy==1.3
progressbar==2.5
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.3
psutil==7.1.1
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pycparser==2.21
pydantic==2.12.3
pydantic-core==2.41.4
pydub==0.25.1
pyerfa==2.0.1.5
pygments==2.19.2
pyparsing==3.2.5
pysocks==1.7.1
pystoi==0.4.1
pytest==8.4.2
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytorchvideo==0.1.5
pywsd==1.2.4
pyyaml==6.0.3
pyzmq==27.1.0
rapidfuzz==3.14.1
regex==2025.10.23
requests==2.32.5
resampy==0.4.3
responses==0.18.0
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rouge-score==0.1.2
rpds-py==0.27.1
s2wrapper==0.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.10.0
send2trash==1.8.3
sentencepiece==0.1.99
sentry-sdk==2.42.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smart-open==7.4.1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
submitit==1.5.3
svgwrite==1.4.3
tabulate==0.9.0
termcolor==3.1.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==0.9.12
tinycss2==1.4.0
tokenizers==0.20.1
tomli==2.3.0
torch==2.8.0+cu128
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.8.0+cu128
torchaudio-augmentations==0.2.1
torchcodec==0.8.0
torchdata==0.7.0
torchdiffeq==0.2.5
torchlibrosa==0.1.0
torchmetrics==1.8.2
torchsde==0.2.6
torchtext==0.16.0
torchvision==0.23.0+cu128
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.46.1
treetable==0.2.6
triton==3.4.0
typeguard==4.4.4
typer==0.20.0
typing-inspection==0.4.2
tyro==0.9.35
tzdata==2025.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.5.0
uv==0.9.4
uvicorn==0.38.0
virtualenv==20.35.3
wandb==0.22.2
wasabi==1.1.3
wavaugment==0.2
wavedrom==2.0.3.post3
wcwidth==0.2.14
weasel==0.4.1
webcolors==24.11.1
webdataset==0.2.86
websocket-client==1.9.0
websockets==15.0.1
wget==3.2
widgetsnbextension==4.0.14
wn==0.13.0
word2number==1.1
wrapt==2.0.0
x-clip==0.14.4
xgrammar==0.1.25
xformers==0.0.32.post2
xgboost==3.0.5
yacs==0.1.8
yarl==1.22.0
zipp==3.23.0
